requirements.txt
streamlit==1.33.0
pandas==2.2.1
plotly==5.20.0
requests==2.31.0
python-dotenv==1.0.1
google-generativeai==0.5.0
streamlit-local-storage==0.0.25
